refactor(engine): Create case management and events service
